Update app.py
Fixes for missing file json_suggestion.crc and other method invocations

+++ seta-data/src/app.py
@@ -42,7 +42,7 @@ def wait_for_es():
         wait_for_es()
 
 
-def suggestion_update_job(config):
+def suggestion_update_job():
     models_path = Config.MODELS_PATH
     if os.path.exists(models_path + Config.WORD2VEC_JSON_EXPORT_CRC):
         crc = open(models_path + Config.WORD2VEC_JSON_EXPORT_CRC, 'r').read()
@@ -53,7 +53,7 @@ def suggestion_update_job(config):
         f.close()
 
     index_suggestion = Config.INDEX_SUGGESTION
-    es = Elasticsearch("http://" + config.ES_HOST, verify_certs=False, request_timeout=30)
+    es = Elasticsearch("http://" + Config.ES_HOST, verify_certs=False, request_timeout=30)
 
     current_w2v_crc, crc_id = get_crc_from_es(es, index_suggestion)
     if crc != current_w2v_crc:
@@ -97,7 +97,7 @@ def delete_all_suggestion(crc):
         print("suggestion delete finished")
 
 
-def seta_es_init_map(config):
+def seta_es_init_map():
     es_session = requests.Session()
     headers = {"Content-Type": "application/json"}
     fn = Config.MODELS_PATH + Config.ES_INIT_DATA_CONFIG_FILE
@@ -134,7 +134,10 @@ def copy_models_files():
 
     dst = Config.MODELS_PATH + Config.WORD2VEC_JSON_EXPORT_CRC
     src = Config.MODELS_DOCKER_PATH + Config.WORD2VEC_JSON_EXPORT_CRC
-    shutil.copyfile(src, dst)
+    
+    #file WORD2VEC_JSON_EXPORT_CRC will be generated in suggestion_update_job()
+    if os.path.exists(src):
+        shutil.copyfile(src, dst)
 
     dst = Config.MODELS_PATH + Config.ES_SUGGESTION_INIT_DATA_CONFIG_FILE
     src = Config.MODELS_DOCKER_PATH + Config.ES_SUGGESTION_INIT_DATA_CONFIG_FILE
@@ -151,17 +154,6 @@ def get_crc_from_es(es, index_suggestion):
         crc_es = resp['hits']['hits'][0]['_source']['crc_model']
         crc_id = resp['hits']['hits'][0]['_id']
     return crc_es, crc_id
-
-
-def getsha256(filename):
-    sha = hash.sha256()
-    with open(filename, 'rb') as f:
-        file_buffer = f.read(BLOCKSIZE)
-        while len(file_buffer) > 0:
-            sha.update(file_buffer)
-            file_buffer = f.read(BLOCKSIZE)
-    return sha.hexdigest()
-
 
 def init():
     wait_for_es()
